requirements.txt
pillow
sv_ttk
pytest
